Move ExecuteOnDispose to Core
